# docs/swagger.yaml
basePath: /api/v1
definitions:
  handler.APIError:
    properties:
      details: {}
      message:
        type: string
    type: object
...
  handler.GenerateMealPlanRequest:
    properties:
      budget:
        type: number
      calorieGoal:
        type: integer
      dietaryPreference:
        type: string
      exclusions:
        items:
          type: string
        type: array
      servings:
        type: integer
      week:
        type: string
    type: object
...
  handler.MealItemResponse:
    properties:
      dayOfWeek:
        type: string
      estimatedCost:
        type: number
      id:
        type: string
      ingredients:
        items:
          type: string
        type: array
      instructions:
        type: string
      mealType:
        type: string
      title:
        type: string
    type: object
  handler.MealPlanResponse:
    properties:
      calorieGoal:
        type: integer
      createdAt:
        type: string
      estimatedCost:
        type: number
      generatedByAi:
        type: boolean
      id:
        type: string
      isoWeek:
        type: string
      items:
        items:
          $ref: '#/definitions/handler.MealItemResponse'
        type: array
    type: object
...
host: localhost:8080
info:
  contact:
    email: contato@example.com
    name: Equipe Golang Finance
  description: API para gestão financeira pessoal com suporte a OCR de recibos.
  title: Golang Finance API
  version: "1.0"
paths:
...
  /meal-plans:
    get:
      description: 'Retorna o plano de refeições salvo para a semana ISO informada
        (padrão: semana atual)'
      parameters:
      - description: 'Semana no formato YYYY-Www (ex: 2024-W37)'
        in: query
        name: week
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.MealPlanResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.APIError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.APIError'
      security:
      - Bearer: []
      summary: Consultar plano de refeições da semana
      tags:
      - Refeições
  /meal-plans/generate:
    post:
      consumes:
      - application/json
      description: Cria um plano semanal com receitas baseadas nos itens de compras
        recentes. Usa heurísticas se o modelo não estiver disponível.
      parameters:
      - description: Preferências para geração
        in: body
        name: body
        schema:
          $ref: '#/definitions/handler.GenerateMealPlanRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.MealPlanResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.APIError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.APIError'
      security:
      - Bearer: []
      summary: Gerar plano de refeições com Gemini
      tags:
      - Refeições
...
schemes:
- http
securityDefinitions:
  Bearer:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"